Revert 82afc03 and fix the underlying issue
VIRTUAL_OBJECT::draw() must not be called with nullptr or invalid
bitmap. So if there is a code path that leads to such a situation,
then it is a bug that has to be fixed.

Add a Theft Bomb missile graphic.
It was missing, which caused the assert() in VIRTUAL_OBJECT::draw()
to go off.

Signed-off-by: Sven Eden <sven@eden-worx.com>